python/requirements.txt
numpy
opencv-python==4.5.4.60
pandas
plotly==4.6.0
pyaml==19.12.0
pytesseract==0.3.8
PyYAML==5.3.1
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.35.0
discord.py
psycopg2-binary
